Continuing refactor. Changing machines again
