+++ client/src/components/dom/Auth/Auth.tsx
@@ -1,38 +1,33 @@
 import * as React from "react";
-import styles from "./Auth.module.scss";
-import AuthModal from "./AuthModal";
-import { useSessionContext } from "@supabase/auth-helpers-react";
+import LoginModal from "./LoginModal";
+import RegisterModal from "./RegisterModal";
+import ConfirmModal from "./ConfirmModal";
+import Nav from "./Nav";
 
 export enum ModalState {
   Login = "Login",
   Register = "Register",
+  Confirm = "Confirm",
   Hidden = "Hidden",
 }
 
 function Auth() {
-  const sessionContext = useSessionContext();
   const [modalState, setModalState] = React.useState<ModalState>(ModalState.Hidden);
-
-  const session = sessionContext.session;
-  const supabase = sessionContext.supabaseClient;
 
   return (
     <>
-      {session === null && (
-        <nav className={styles.nav}>
-          <button onClick={() => setModalState(ModalState.Register)}>Register</button>
-          <button onClick={() => setModalState(ModalState.Login)}>Sign In</button>
-        </nav>
+      <Nav modalState={modalState} setModalState={setModalState} />
+
+      {modalState === ModalState.Login && (
+        <LoginModal modalState={modalState} setModalState={setModalState} />
       )}
-      {session !== null && (
-        <nav className={styles.nav}>
-          <p>{session.user.email}</p>
-          <button onClick={() => supabase.auth.signOut()}>Sign Out</button>
-        </nav>
+      {modalState === ModalState.Register && (
+        <RegisterModal modalState={modalState} setModalState={setModalState} />
       )}
-      {modalState !== ModalState.Hidden && (
-        <AuthModal modalState={modalState} setModalState={setModalState} />
+      {modalState === ModalState.Confirm && (
+        <ConfirmModal modalState={modalState} setModalState={setModalState} />
       )}
+      {modalState === ModalState.Hidden && <></>}
     </>
   );
 }
