+++ client/src/components/dom/Auth/ConfirmModal/ConfirmModal.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+import { ModalState } from "../Auth";
+
+export default function ConfirmModal(props: {
+  modalState: ModalState;
+  setModalState: (state: ModalState) => void;
+}) {
+  return <></>;
+}
