+++ client/src/components/dom/Auth/ConfirmModal/index.ts
@@ -1,0 +1,3 @@
+import ComfirmModal from './ConfirmModal';
+
+export default ComfirmModal;
