+++ client/src/components/dom/Auth/Form/Form.tsx
@@ -1,0 +1,3 @@
+export default function Form() {
+  return <></>;
+}
