+++ client/src/components/dom/Auth/Nav/Nav.module.scss
@@ -1,0 +1,27 @@
+.nav {
+  position: fixed;
+  z-index: 20;
+  top: 0;
+  right: 0;
+
+  display: flex;
+  justify-content: flex-end;
+  gap: 2rem;
+
+  button {
+    background-color: rgba(0, 0, 0, 0);
+    cursor: pointer;
+    padding: 1em 2em;
+
+    border: none;
+    text-decoration: none;
+    color: white;
+    font-size: 1.2em;
+  }
+
+  p {
+    padding: 1em 2em;
+    color: white;
+    font-size: 1.2em;
+  }
+}
