+++ client/src/components/dom/Auth/Nav/index.ts
@@ -1,0 +1,3 @@
+import Nav from "./Nav";
+
+export default Nav;
