+++ client/src/components/dom/Auth/RegisterModal/RegisterModal.tsx
@@ -1,33 +1,35 @@
 import * as React from "react";
 import { useSupabaseClient } from "@supabase/auth-helpers-react";
 import styles from "../AuthModal.module.scss";
-import { ModalState } from "../../Auth";
+import { ModalState } from "../Auth";
 
 type FormDetails = {
   email: string;
   password: string;
+  confirm: string;
 };
 type FormErrors = {
   email: string;
   password: string;
+  confirm: string;
 };
 
-export default function LoginModal(props: {
+export default function RegisterModal(props: {
   modalState: ModalState;
-  setModalState: (action: ModalState) => void;
+  setModalState: (state: ModalState) => void;
 }) {
   const supabase = useSupabaseClient();
-
   /*COMPONENT STATE*/
   const [formDetails, setFormDetails] = React.useState<FormDetails>({
     email: "",
     password: "",
+    confirm: "",
   });
   const [formErrors, setFormErrors] = React.useState<FormErrors>({
     email: "",
     password: "",
+    confirm: "",
   });
-
   /*COMPONENT FUNCTIONS*/
 
   function handleInputChange(e: any) {
@@ -35,10 +37,11 @@ export default function LoginModal(props: {
   }
 
   function resetForm() {
-    setFormDetails({ email: "", password: "" });
+    setFormDetails({ email: "", password: "", confirm: "" });
   }
+
   function resetErrors() {
-    setFormErrors({ email: "", password: "" });
+    setFormErrors({ email: "", password: "", confirm: "" });
   }
 
   function closeModal(e: any) {
@@ -52,45 +55,58 @@ export default function LoginModal(props: {
   function switchActions() {
     resetForm();
     resetErrors();
-    props.setModalState(ModalState.Register);
+    props.setModalState(ModalState.Login);
   }
 
-  function verifyLoginInputs() {
+  function verifyRegisterInputs() {
     let isValid = true;
     resetErrors();
 
+    // Check password length
+    if (formDetails.password.length < 6) {
+      isValid = false;
+      setFormErrors(prev => ({ ...prev, password: "Password must be at least 6 characters" }));
+    }
+    // Check if passwords match
+    if (formDetails.password !== formDetails.confirm) {
+      isValid = false;
+      setFormErrors(prev => ({ ...prev, confirm: "Passwords do not match" }));
+    }
     // Check if email is valid
     if (!formDetails.email.includes("@")) {
       isValid = false;
       setFormErrors(prev => ({ ...prev, email: "Invalid email" }));
     }
-    if (formDetails.password === "") {
-      isValid = false;
-      setFormErrors(prev => ({ ...prev, password: "Required" }));
+    // Check if any fields are empty
+    for (const key in formDetails) {
+      if (formDetails[key] === "") {
+        isValid = false;
+        setFormErrors(prev => ({ ...prev, [key]: "Required" }));
+      }
     }
     return isValid;
   }
 
-  async function handleLogin() {
+  async function handleRegister() {
     // Check if inputs are valid
-    if (!verifyLoginInputs()) return;
+    if (!verifyRegisterInputs()) return;
 
     // Send request to supabase
-    const { error } = await supabase.auth.signInWithPassword({
+    const { data, error } = await supabase.auth.signUp({
       email: formDetails.email,
       password: formDetails.password,
     });
     // Check for api errors
     if (error) {
       // console.log(error);
-
-      setFormErrors(prev => ({
-        ...prev,
-        email: "Invalid email or password",
-        password: "Invalid email or password",
-      }));
       return;
     }
+    // Check if account already exists
+    if (!data.user.identities.length) {
+      setFormErrors(prev => ({ ...prev, email: "Account already exists" }));
+      return;
+    }
+    // Account created, needs email verification
     resetForm();
     resetErrors();
     props.setModalState(ModalState.Hidden);
@@ -99,7 +115,7 @@ export default function LoginModal(props: {
   return (
     <div className={styles.authModal} onClick={closeModal}>
       <div className={styles.content}>
-        <h1>Login</h1>
+        <h1>Register</h1>
         <div className={styles.inputSection}>
           <label htmlFor="email">Email</label>
           <input
@@ -114,9 +130,6 @@ export default function LoginModal(props: {
         <div className={styles.inputSection}>
           <div className={styles.passwordLabels}>
             <label htmlFor="password">Password</label>
-            <a href="#" className="forgotPassword">
-              Forgot Password
-            </a>
           </div>
           <input
             type="password"
@@ -127,15 +140,26 @@ export default function LoginModal(props: {
           />
           {formErrors.password && <p className={styles.error}>{formErrors.password}</p>}
         </div>
+        <div className={styles.inputSection}>
+          <label htmlFor="confirm">Confirm Password</label>
+          <input
+            type="password"
+            name="confirm"
+            id="confirm"
+            value={formDetails.confirm}
+            onChange={handleInputChange}
+          />
+          {formErrors.confirm && <p className={styles.error}>{formErrors.confirm}</p>}
+        </div>
         <div className={styles.buttonWrapper}>
           <button onClick={closeModal}>Cancel</button>
-          <button className={styles.actionButton} onClick={handleLogin}>
-            Sign in
+          <button className={styles.actionButton} onClick={handleRegister}>
+            Sign Up
           </button>
         </div>
         {/* Option to switch between login and register */}
         <button onClick={switchActions} className={styles.switchActions}>
-          Don't have an account?
+          Already have an account?
         </button>
       </div>
     </div>
