+++ client/src/components/dom/Auth/RegisterModal/index.ts
